backend/app/requirements.txt
addict==2.4.0
aioredis==1.3.1
emails==0.6
fastapi==0.66.0
fastapi_cache==0.1.0
fastapi-cache2
fastapi_profiler==1.0.0
gevent_socketio==0.3.6
ldap3==2.9.1
loguru==0.5.3
openpyxl==3.0.7
pandas==1.4.4
paramiko==2.7.2
passlib==1.7.4
psutil==5.9.0
pycryptodome==3.16.0
pydantic==1.10.2
pymongo==3.11.4
pytest==7.2.0
python-dotenv==0.21.0
python_jose==3.3.0
pytz==2022.2.1
regex==2022.9.13
slowapi==0.1.7
SQLAlchemy==1.3.17
starlette==0.14.2
tenacity==8.1.0
tomlkit==0.11.6
uvicorn==0.14.0
yapf==0.26.0
pymysql
redis
pydantic[email]
python-multipart
aiofiles
Jinja2==3.0.1
